Users/appleptr16/first tables

+++ src/main/java/com/ambrosia/loans/discord/base/Emeralds.java
@@ -43,7 +43,7 @@ public class Emeralds {
     }
 
 
-    public static int leToEmeralds(int le) {
-        return LIQUID * le;
+    public static int leToEmeralds(double le) {
+        return (int)(LIQUID * le);
     }
 }
